pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "BiteAWX"
description = "For reading AWX format data."
authors = [
    {name = "modabao", email = "mo_dabao@qq.com"}
]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3"
]
dependencies = [
    "numpy",
    "xarray",
    "pyproj",
    "pillow",
    "netCDF4"
]
requires-python = ">=3.6"
dynamic = ["version", "readme"]

[tool.setuptools.dynamic]
version = {attr = "BiteAWX.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.setuptools]
packages = ["BiteAWX"]

[project.urls]
repository = "https://github.com/Mo-Dabao/BiteAWX.git"